requirements.txt
streamlit
opencv-python
pandas
numpy
pillow
pytesseract
speechrecognition
pyttsx3
langchain
chromadb
openai
matplotlib
plotly
mediapipe
PyPDF2
python-docx
librosa
whisper
faiss-cpu
sentence-transformers
transformers
torch
torchvision
scikit-learn
seaborn
pyaudio
webrtcvad
streamlit-webrtc
streamlit-option-menu
streamlit-lottie
requests
networkx
igraph
pyvis
spacy
nltk
textblob
wordcloud
bokeh
altair
folium
streamlit-folium
streamlit-aggrid
streamlit-camera-input-live
streamlit-drawable-canvas
streamlit-ace
streamlit-chat
streamlit-elements
streamlit-player
streamlit-image-coordinates
streamlit-cropper
streamlit-toggle-switch
streamlit-card
streamlit-metrics
streamlit-timeline
streamlit-flow-component
streamlit-3d-viewer
jupyter
ipython
sympy
scipy
statsmodels
plotly-dash
dash
flask
fastapi
uvicorn
sqlalchemy
sqlite3
psycopg2-binary
mysql-connector-python
redis
celery
pytest
black
flake8
mypy
pre-commit
docker
kubernetes
tensorflow
keras
pytorch-lightning
huggingface-hub
datasets
accelerate
gradio
wandb
mlflow
dvc
great-expectations
evidently
streamlit-authenticator
streamlit-extras
streamlit-shadcn-ui
streamlit-antd-components
